// shared/src/models/User.ts
import { ObjectId } from "mongodb";
export interface User {
  _partitionKey: string;

  _id: ObjectId;

  userType: string;

  userId: string;

  firstName: string;

  lastName: string;

  gender: string;

  countryId: string;

  associationId: string;

  associationName: string;

  fcmToken: string;

  email: string;

  cellphone: string;

  password: string;

  countryName: string;

  dateRegistered: string;

  qrCodeUrl: string;

  profileUrl: string;

  profileThumbnail: string;
}